% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livercells.R
\name{livercells}
\alias{livercells}
\title{livercells is A package for deconvolute the liver RNA-seq data for cell composition}
\usage{
livercells(
  expr,
  signatures = NULL,
  genes = NULL,
  parallel.sz = 4,
  parallel.type = "SOCK"
)
}
\arguments{
\item{expr}{is the gene expression data set. A matrix with rownames as gene symbol and colmanes as samples}

\item{signatures}{a GMT object of signatures.}

\item{genes}{list of genes to use in the analysis.}

\item{parallel.sz}{integer for the number of threads to use. Default is 4.}

\item{parallel.type}{Type of cluster architecture when using snow. 'SOCK' or 'FORK'. Fork is faster, but is not supported in windows.}
}
\value{
dataframe with enrichmnet score for each cells in every samples
}
\description{
This function  provide the cell composition in any liver tissue
}
\author{
Rama Shankar
}
